smart-store-agent/pyproject.toml
[project]
name = "smart-store-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
]
